Simplify video player cards
